chore: merge test files
